Added multylink support

+++ MainWindow.xaml.cs
@@ -16,6 +16,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Net;
+using System.Net.Sockets;
 using System.Threading;
 using System.Windows;
 using System.Windows.Controls;
@@ -31,21 +32,22 @@ namespace YandereSimLauncher {
 
         private const string BASE_LINK =    "http://yanderesimulator.com/";
         private const string URLS_LINK =    "http://yanderesimulator.com/urls.txt";
+        //private const string URLS_LINK =    "http://localhost:8080/urls.txt";
         private const string NEWS_URL =     "https://public-api.wordpress.com/rest/v1.1/sites/yanderedev.wordpress.com/posts/";
         private const string ZIP_NAME =     "content.zip";
-        private const int VERSION =         1;
+        private const int VERSION =         2;
 
         private enum GameStatus { Updated, Outdated, NotDownloaded, ContentError }
 
         private enum LinkType {
-            version, contents, newlauncher, launcher,
+            version, newlauncher, launcher,
             wordpress, youtube, twitter, twitch,
             volunteer, contact, about, donate
         }
 
         private Dictionary<LinkType, string> Links = new Dictionary<LinkType, string>() {
             { LinkType.version,     "http://yanderesimulator.com/version.txt" },
-            { LinkType.contents,    "http://yanderesimulator.com/latest.zip" },
+            //{ LinkType.contents,    "http://yanderesimulator.com/latest.zip" },
             { LinkType.launcher,    "http://yanderesimulator.com/launcherversion.txt" },
             { LinkType.newlauncher, "http://yanderesimulator.com/download/" },
 
@@ -60,9 +62,12 @@ namespace YandereSimLauncher {
             { LinkType.donate,      "http://yanderesimulator.com/donate/" },
         };
 
+        private List<string> contentLinks = new List<string>();
+
         private WebClient webClient;
         private string gamePath;
         private int newGameVersion;
+        private int curLink;
         private Thread launcherThread;
         private bool isAppClosed;
         private bool isLinkUpdated;
@@ -247,20 +252,37 @@ namespace YandereSimLauncher {
         }
 
         private void CleanUp() {
-            var executable = Directory.EnumerateFiles(gamePath, "YandereSimulator.exe");
-            var content = Directory.EnumerateDirectories(gamePath, "YandereSimulator_Data");
-            if (executable.Any()) File.Delete(executable.ElementAt(0));
-            if (content.Any()) DeleteDirectory(content.ElementAt(0));
+            try {
+                var executable = Directory.EnumerateFiles(gamePath, "YandereSimulator.exe");
+                var content = Directory.EnumerateDirectories(gamePath, "YandereSimulator_Data");
+                if (executable.Any()) File.Delete(executable.ElementAt(0));
+                if (content.Any()) DeleteDirectory(content.ElementAt(0));
+            } catch (IOException) {
+                MessageBox.Show("Looks like you're running Yandere Simulator. Please close the game and try again.", "Ooops!");
+                Dispatcher.Invoke(new Action(() => { Close(); }));
+            }
         }
 
         private void DownloadGameContents() {
-            ReportStatus("Starting download");
-            CleanUp();
+            if (contentLinks.Count > curLink) {
+                ReportStatus("Starting download");
+                CleanUp();
 
-            if (File.Exists(gamePath + ZIP_NAME)) File.Delete(gamePath + ZIP_NAME);
-            webClient.DownloadProgressChanged += new DownloadProgressChangedEventHandler(UpdateProgressBar);
-            webClient.DownloadFileCompleted += new AsyncCompletedEventHandler(DownloadDataCompleted);
-            webClient.DownloadFileAsync(new Uri(Links[LinkType.contents]), gamePath + ZIP_NAME);
+                if (File.Exists(gamePath + ZIP_NAME)) File.Delete(gamePath + ZIP_NAME);
+                try {
+                    webClient.DownloadProgressChanged += new DownloadProgressChangedEventHandler(UpdateProgressBar);
+                    webClient.DownloadFileCompleted += new AsyncCompletedEventHandler(DownloadDataCompleted);
+                    webClient.DownloadFileAsync(new Uri(contentLinks[curLink]), gamePath + ZIP_NAME);
+                } catch (SocketException) {
+                    curLink++;
+                    if (launcherThread != null & launcherThread.IsAlive) launcherThread.Abort();
+                    launcherThread = new Thread(StartCheckingInternetConnection);
+                    launcherThread.Start();
+                }
+            } else {
+                SetServerUnavailableStatus("Can't reach download source");
+                MessageBox.Show("Can't reach download source. Maybe it is unavailable now or your antivirus blocks the connection", "Ooops!");
+            }
         }
 
         private void DownloadDataCompleted(object sender, AsyncCompletedEventArgs e) {
@@ -276,12 +298,14 @@ namespace YandereSimLauncher {
                 }
                 File.Delete(gamePath + ZIP_NAME);
             } catch {
-                ReportStatus("Can't finish extracting");
-                Dispatcher.Invoke(new Action(() => {
-                    RedownloadButton.IsEnabled = true;
-                    PlayButton.IsEnabled = false;
-                }));
+                ReportStatus("Error. Retrying...");
+                curLink++;
+                if (launcherThread != null & launcherThread.IsAlive) launcherThread.Abort();
+                launcherThread = new Thread(StartCheckingInternetConnection);
+                launcherThread.Start();
+                return;
             }
+
             try {
                 File.WriteAllText(gamePath + "YandereSimulator_Data\\" + "version", newGameVersion.ToString());
                 SetReadyStatus();
@@ -323,21 +347,24 @@ namespace YandereSimLauncher {
                     try {
                         var lnk = l.Split('"');
 
-                        var newLink = lnk[1].Trim();
-                        if (!newLink.StartsWith("http")) {
-                            newLink = "http://" + newLink;
-                        }
+                        if (lnk[0].Trim().StartsWith("contents")) {
+                            contentLinks.Add(GetLink(lnk[1]));
 
-                        Links[GetLinkType(lnk[0])] = newLink;
+                        } else Links[GetLinkType(lnk[0])] = GetLink(lnk[1]);
 
-                    } catch (Exception) {
-                        //Really don't care what shit happened, just
-                        continue;
-                    }
+                    } catch (Exception) { continue; }
                 }
                 isLinkUpdated = true;
 
             } catch (WebException) { }
+        }
+
+        private string GetLink(string inp) {
+            var newLink = inp.Trim();
+            if (!newLink.StartsWith("http")) {
+                newLink = "http://" + newLink;
+            }
+            return newLink;
         }
 
         private bool IsLauncherUpdated() {
@@ -362,24 +389,28 @@ namespace YandereSimLauncher {
                 Thread t = new Thread(StartCheckThread);
                 t.Start();
             } catch (WebException) {
-                ReportStatus("Can't connect to update server");
-                try {
-                    var executable = Directory.EnumerateFiles(gamePath, "YandereSimulator.exe").ElementAt(0);
-                    var content = Directory.EnumerateDirectories(gamePath, "YandereSimulator_Data").ElementAt(0);
-                    Dispatcher.Invoke(new Action(() => {
-                        PlayButton.IsEnabled = true;
-                    }));
-                } catch {
-                    Dispatcher.Invoke(new Action(() => {
-                        RedownloadButton.IsEnabled = false;
-                        PlayButton.IsEnabled = false;
-                    }));
-                }
+                SetServerUnavailableStatus("Can't connect to update server");
                 Thread.Sleep(5000);
                 if (!isAppClosed) {
                     launcherThread = new Thread(StartCheckingInternetConnection);
                     launcherThread.Start();
                 }
+            } 
+        }
+
+        private void SetServerUnavailableStatus(string serverStatus) {
+            ReportStatus(serverStatus);
+            try {
+                var executable = Directory.EnumerateFiles(gamePath, "YandereSimulator.exe").ElementAt(0);
+                var content = Directory.EnumerateDirectories(gamePath, "YandereSimulator_Data").ElementAt(0);
+                Dispatcher.Invoke(new Action(() => {
+                    PlayButton.IsEnabled = true;
+                }));
+            } catch {
+                Dispatcher.Invoke(new Action(() => {
+                    RedownloadButton.IsEnabled = false;
+                    PlayButton.IsEnabled = false;
+                }));
             }
         }
